Revert "Refactor component structure: move Header and Footer to commons directory, reintroduce SearchInput component"
This reverts commit d554c918e0b3e43bd40cdfe6ee1fc5b923458bd5.
